+++ packages/project-access/src/index.ts
@@ -5,7 +5,6 @@ export {
     clearCdsModuleCache,
     createApplicationAccess,
     createProjectAccess,
-    isPathForCapApp,
     deleteCapApp,
     filterDataSourcesByType,
     findAllApps,