+++ packages/ui5-library-reference-writer/CHANGELOG.md
@@ -1,5 +1,26 @@
 # @sap-ux/ui5-library-reference-writer
 
+## 0.1.30
+
+### Patch Changes
+
+-   Updated dependencies [eb38e5b]
+    -   @sap-ux/project-access@1.28.2
+
+## 0.1.29
+
+### Patch Changes
+
+-   Updated dependencies [64e037d]
+    -   @sap-ux/project-access@1.28.1
+
+## 0.1.28
+
+### Patch Changes
+
+-   Updated dependencies [15e6959]
+    -   @sap-ux/project-access@1.28.0
+
 ## 0.1.27
 
 ### Patch Changes